note key
I need help completing this Unity music visualization tool. The coloring and positioning of the Umbilic torus that visualizes the circle of fifths, depends on the key of the music. For now, let's set the key manually with a user input (a-b-c-d-e-f-g keys on my computer keyboard - ignore sharps). the umbilic torus needs to "twist" and rotate when the key is changed. If you understand the geometry of the torus, you will understand how to do this. The key should be at the inner 12 o'clock position as it is currently designed, with the major 5th counterclockwise to it, and the major 3rd clockwise to it, as it is currently designed

The coloring of the system is based on the following rules:
1 - the key is blue, the fifth is green, and the 3rd is red
2 - the minors are the blend of each of these colors depending on what notes they share with their major counterparts
3 - the individual notes are the blends of the blends. if a note is played by itself (i.e. C in the key of C, it is pure blue. if it is a part of a chord, it is blended with that chord color)
4 - I included a yellow color at either end of the spectrum, for the color to wrap around on itself at the Neapolitan

the color/interval correlation is hard coded and should be respected as it is currently written. the note color calculation needs to also read the input notes at that moment in time in the same way that the chords are colored, and colored correctly according to the rules above

fully implement these two changes:
1 - key change and twist rotating the toroidal geometry
2 - coloring individual notes based on the relative other notes that are playing